Move block break animation broadcasting to GlowPlayer#setDigging()
