+++ src/main/java/net/glowstone/entity/GlowPlayer.java
@@ -2838,8 +2838,22 @@ public class GlowPlayer extends GlowHumanEntity implements Player {
         return digging;
     }
 
+    private void broadcastBlockBreakAnimation(GlowBlock block, int destroyStage) {
+        if (block == null) return;
+
+        GlowChunk.Key key = new GlowChunk.Key(block.getChunk().getX(), block.getChunk().getZ());
+        // TODO: should canSeeChunk filtering take place here, in GlowPlayer, or both?
+        block.getWorld().getRawPlayers().stream().filter(player1 -> player1.canSeeChunk(key)).forEach(player2 -> {
+            player2.sendBlockBreakAnimation(player2, block.getLocation(), destroyStage);
+        });
+    }
+
     public void setDigging(GlowBlock block) {
         digging = block;
+
+        // show other clients the block is beginning to crack
+        // TODO: show incremental stages
+        broadcastBlockBreakAnimation(block, 0);
     }
 
     @Override
